+++ src/main/java/com/server/EZY/exception/authentication_number/controller/AuthenticationNumberExceptionController.java
@@ -1,7 +1,7 @@
-package com.server.EZY.exception.authenticationNumber.controller;
+package com.server.EZY.exception.authentication_number.controller;
 
-import com.server.EZY.exception.authenticationNumber.exception.AuthenticationNumberTransferFailedException;
-import com.server.EZY.exception.authenticationNumber.exception.InvalidAuthenticationNumberException;
+import com.server.EZY.exception.authentication_number.exception.AuthenticationNumberTransferFailedException;
+import com.server.EZY.exception.authentication_number.exception.InvalidAuthenticationNumberException;
 import com.server.EZY.response.result.CommonResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
