+++ src/main/java/com/server/EZY/exception/authentication_number/exception/InvalidAuthenticationNumberException.java
@@ -1,4 +1,4 @@
-package com.server.EZY.exception.authenticationNumber.exception;
+package com.server.EZY.exception.authentication_number.exception;
 
 public class InvalidAuthenticationNumberException extends RuntimeException{
     public InvalidAuthenticationNumberException(String msg, Throwable t){
