+++ src/main/java/com/server/EZY/exception/invalidRequest_data/InvalidRequestDataHandler.java
@@ -1,4 +1,4 @@
-package com.server.EZY.exception.invalidRequestData;
+package com.server.EZY.exception.invalidRequest_data;
 
 import com.server.EZY.response.ResponseService;
 import com.server.EZY.response.result.CommonResult;
@@ -13,10 +13,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 @Slf4j
 @RestControllerAdvice @Order(Ordered.HIGHEST_PRECEDENCE)
