+++ src/main/java/com/server/EZY/security/authentication/MyUserDetails.java
@@ -1,4 +1,4 @@
-package com.server.EZY.security.Authentication;
+package com.server.EZY.security.authentication;
 
 import com.server.EZY.model.member.MemberEntity;
 import com.server.EZY.model.member.repository.MemberRepository;
