+++ src/test/java/com/server/EZY/security/authentication/MyUserDetailsTest.java
@@ -1,10 +1,9 @@
-package com.server.EZY.security.Authentication;
+package com.server.EZY.security.authentication;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
 import static org.junit.jupiter.api.Assertions.*;
